Revert "Refactor functions"
