+++ lib/mariadb.class.js
@@ -4,19 +4,14 @@ const crypto = require('crypto');
 'use strict';
 
 /**
- * @ignore
- * @param {string} msg - string to be hashed
- * @returns {string} Hashed message
+ *	@param {string} msg - string to be hashed
+ *	@returns {string} Hashed message
  * */
 function SHA256(msg) {
 	return crypto.createHash('sha256').update(msg).digest('hex');
 }
 
-/**
- * mariadb helper
- * @class
- * */
-class DATABASE {
+class Database {
 	constructor(config, rejectEmpty = false, limitQueryExecution = true) {
 		this._config = config;
 		this._rejectEmpty = config.rejectEmpty || rejectEmpty;
@@ -31,8 +26,8 @@ class DATABASE {
 	};
 
 	/**
-	 * Expose mariadb package
-	 * @returns {Object} current connection
+	 *	Expose mariadb package
+	 *	@returns {Object} current connection
 	 * */
 	connection() {
 		return __sqlPools[this._connectionHash]
@@ -41,9 +36,9 @@ class DATABASE {
 	};
 
 	/**
-	 * Escape undefined in args as null. Preventing query execution from throwing error
-	 * @param {string[]} args - arguments to be passed into query
-	 * @returns {string[]} escaped args
+	 *	Escape undefined in args as null. Preventing query execution from throwing error
+	 * 	@param {string[]} args - arguments to be passed into query
+	 * 	@returns {string[]} escaped args
 	 * */
 	escapeUndefined(args){
 		if(!args instanceof Object){
@@ -60,12 +55,11 @@ class DATABASE {
 	};
 
 	/**
-	 * Execute query with arguments
-	 * @async
-	 * @param {string} sql - sql query
-	 * @param {string[]} args - escaped arguments to be passed into query (avoiding injection)
-	 * @param {boolean} [dateStrings=true] - if false datetime columns will be returned as js Date object
-	 * @returns {Object[]} sql query result
+	 *	Execute query with arguments
+	 * 	@param {string} sql - sql query
+	 * 	@param {string[]} args - escaped arguments to be passed into query (avoiding injection)
+	 * 	@param {boolean} [dateStrings=true] - if false datetime columns will be returned as js Date object
+	 * 	@returns {Object[]} sql query result
 	 * */
 	query(sql, args =[], stripMeta = false, dateStrings = true) {
 		return new Promise(async (resolve, reject) => {
@@ -109,12 +103,11 @@ class DATABASE {
 	};
 
 	/**
-	 * Execute query batch with arguments
-	 * @async
-	 * @param {string} sql - sql query
-	 * @param {string[]} args - escaped arguments to be passed into query (avoiding injection)
-	 * @param {boolean} [dateStrings=true] - if false datetime columns will be returned as js Date object
-	 * @returns {Object[]} sql query result
+	 *	Execute query batch with arguments
+	 * 	@param {string} sql - sql query
+	 * 	@param {string[]} args - escaped arguments to be passed into query (avoiding injection)
+	 * 	@param {boolean} [dateStrings=true] - if false datetime columns will be returned as js Date object
+	 * 	@returns {Object[]} sql query result
 	 * */
 	batch(sql, args, dateStrings = true) {
 		return new Promise(async (resolve, reject) => {
@@ -156,10 +149,10 @@ class DATABASE {
 	};
 
 	/**
-	 * Debug SQL query with arguments
-	 * @param {string} sql - sql query
-	 * @param {string[]} args - escaped arguments to be passed into query (avoiding injection)
-	 * @returns {Object[]} sql query with arguments
+	 *	Debug SQL query with arguments
+	 * 	@param {string} sql - sql query
+	 * 	@param {string[]} args - escaped arguments to be passed into query (avoiding injection)
+	 * 	@returns {Object[]} sql query with arguments
 	 * */
 	debug(sql, args) {
 		if(!Array.isArray(args)){
@@ -178,7 +171,7 @@ class DATABASE {
 	}
 
 	/**
-	 * End current connection and delete it from global object
+	 *	End current connection and delete it from global object
 	 * */
 	async end() {
 		await __sqlPools[this._connectionHash].end();
@@ -186,4 +179,4 @@ class DATABASE {
 	};
 }
 
-module.exports = DATABASE;
+module.exports = Database;
