+++ lib/sqlImporter.class.js
@@ -1,35 +1,5 @@
-/**
- * @file sqlImporter.class.js
- * */
 const fs = require('fs');
 const _mariadb = require('./mariadb.class.js');
-
-const {
-	getDelimiters,
-	escapeRegex,
-	removeDelimiters,
-	changeDefiner,
-	checkAlterHasPrimary,
-	alterMultipleToSingle,
-	insertMultipleToSingle,
-	getColumnsFromCreateTable,
-	parseQueryRoutine,
-	getNameFromQuery,
-	getNameFromQueries,
-	getForeignKeyRelationship,
-	getTableDependencies,
-	topologySortDependencies,
-	parseColumnCreateQuery,
-	stripColumnDefinitionFromCreate,
-	getNonTableCreates,
-	isColumnIdentical,
-	parseSQLInput,
-} = require('./functions.lib.js');
-
-/**
- * @defgroup SQLImporter SQL Importer
- * @brief Import SQL database from a file.
- * */
 
 /**
  * Error from process.
@@ -42,71 +12,40 @@ const {
  */
 
 /**
- * Failed Query
- * @typedef {Object} FailedQuery
- * @property {number} code - SQL Error number
- * @property {string} query - Failed SQL Query
- */
-
-/**
- * Colum Comparation detail
- * @typedef {Object} ColCompDetail
- * @property {string} name - column name
- * @property {string} detail - column detail, i.e. 'INT NOT NULL'
- */
-
-/**
- * Table Comparation result
- * @typedef {Object} TableCompare
- * @property {ColCompDetail[]} new - new columns in the table
- * @property {ColCompDetail[]} same - same columns in the table
- * @property {ColCompDetail[]} mod - modified columns in the table
- * @property {String[]} nomore - SQL Error number
- */
-
-/**
  * Response object with status, data and error
- * @ingroup SQLImporter
  * @typedef {Object} Response
  * @property {boolean} status - indicates whether the process run successfully or not.
  * @property {resError|undefined} error - presents if status is false, otherwise it's undefined.
  * @property {resData|undefined} data - presents if status is true, otherwise it's undefined.
  */
+
 function response(status, content, debug = false){
 	status = Boolean(status);
 
-	if(debug) status ? console.log(content) : console.error(content);
+	if(debug){
+		status ? console.log(content) : console.error(content);
+	}
 
 	return status
 		? {status, data: content}
 		: {status, error: content};
 }
-/**
- * @ingroup SQLImporter
- * @class SQL_IMPORTER lib/sqlImporter.class.js
- */
-class SQL_IMPORTER {
-	/**
-	 * @var DATABASE _mysql
-	*/
-	_mysql = undefined;
 
-	/**
-	 * @private
-	 * */
+class IMPORTER {
 	constructor(config){
 		const self = this;
 
 		self._config = undefined;
-
+		self._mysql = undefined;
 		self._contents = undefined;
 
-		if(config) self.init(config);
+		if(config){
+			self.init(config);
+		}
 	}
 
 	/**
 	 * Initiatlize instance with config
-	 *
 	 * @param {Object} config - mysql connection config
 	 * @returns {Object} class instance
 	 * */
@@ -123,8 +62,6 @@ class SQL_IMPORTER {
 
 	/**
 	 * close and delete mysql connection
-	 *
-	 * @async
 	 * */
 	async close(){
 		const self = this;
@@ -137,13 +74,10 @@ class SQL_IMPORTER {
 
 	/**
 	 * Get all tables from connected database
-	 *
-	 * @private
-	 * @async
 	 * @returns {Response} response - response from mysql query
 	 * @returns {string[]} response.data - array of tables' names
 	 * */
-	_getAllTables(){
+	getAllTables(){
 		const self = this;
 
 		return self._mysql.query(`SHOW TABLES`)
@@ -164,12 +98,9 @@ class SQL_IMPORTER {
 
 	/**
 	 * Drop table from connected database
-	 *
-	 * @private
-	 * @async
-	 * @param {string} name - table's name
+	 * @params {string} name - table's name
 	 * */
-	_dropTable(name){
+	dropTable(name){
 		const self = this;
 
 		const { verbose } = self._config;
@@ -205,39 +136,33 @@ class SQL_IMPORTER {
 
 	/**
 	 * Drop all tables from connected database
-	 *
-	 * @private
-	 * @param database **[string]** Database name
 	 * @returns {Response|undefined} response - will return response if error is encountered
 	 * */
-	async _dropAllTables(database){
+	async dropAllTables(database){
 		const self = this;
 
-		const tables = await self._getAllTables();
+		const tables = await self.getAllTables();
 
 		if(!tables.status){
 			return response(true, 'Db contains no Tables');
 		}
 
 		for(const name of tables.data){
-			await self._dropTable(name);
+			await self.dropTable(name);
 		}
-	}
+	};
 
 	/**
 	 * Drop all routines from connected database
-	 *
-	 * @private
-	 * @param database **[string]** Database name
 	 * @returns {Response|undefined} response - will return response if error is encountered
 	 * */
-	async _dropRoutines(database){
+	async dropRoutines(database){
 		const self = this;
 
 		const { verbose } = self._config;
 
 		for(const rName of [ 'FUNCTION', 'PROCEDURE' ]){
-			const rRoutines = await self._getAllRoutines(database, rName);
+			const rRoutines = await self.getAllRoutines(database, rName);
 
 			if(!rRoutines.status){
 				return response(true, `Db contains no '${rName}'`);
@@ -275,23 +200,20 @@ class SQL_IMPORTER {
 
 	/**
 	 * Get all routines from connected database
-	 *
-	 * @private
-	 * @param database **[string]** Database name
 	 * @returns {Response} response
 	 * @returns {string[]} response.data - array of routines' names
 	 * */
-	async _getAllRoutines(database, rName){
+	async getAllRoutines(database, rName){
 		const self = this;
 
 		rName = rName.toUpperCase();
 
-		const _sql = {
+		const sql = {
 			query: `SHOW ${rName} STATUS WHERE Db = ?`,
-			params: [ database ],
+			params: [database],
 		};
 
-		return self._mysql.query(_sql.query, _sql.params)
+		return self._mysql.query(sql.query, sql.params)
 			.then(res => {
 				const data = [];
 				for(const item of res){
@@ -306,10 +228,352 @@ class SQL_IMPORTER {
 	}
 
 	/**
-	 * Remove tables and routines from connected database
+	 * Change definer's name and host according to configuration
+	 * @params {string} query - query string conatining DEFINER clause
+	 * @returns {string} query string with modified DEFINER clause
+	 * */
+	_changeDefiner(query){
+		const self = this;
+
+		return query.replace(
+			/DEFINER=(.*?)\@(.*?)\s/,
+			'DEFINER=`' + self._config.user + '`'
+					+ '@`' + self._config.host + '` '
+		);
+	};
+
+	/**
+	 * Get all CREATE clauses for not table, e.g. FUNCTION
+	 * @params {string} sql - entire sql string from file
+	 * @returns {Object} results - CREATE queries
+	 * @returns {string[]} results.functions - FUNCTION's CREATE queries
+	 * @returns {string[]} results.procedures - PROCEDURE's CREATE queries
+	 * @returns {string[]} results.triggers - TRIGGER's CREATE queries
+	 * */
+	getNonTableCreates(sql){
+		const self = this;
+
+		const results = {
+			functions: [],
+			procedures: [],
+			triggers: [],
+		};
+
+		const getDelimiters = function(queries){
+			let _delimiters = queries.match(/DELIMITER\s([\S]+)/gm);
+			const delimiters = [];
+
+			if(_delimiters){
+				_delimiters = Array.from(new Set(_delimiters));
+				const length = _delimiters.length;
+
+				for(let idx = 0; idx < length; idx++){
+					const delimiter = _delimiters[idx]
+						.replace('DELIMITER', '')
+						.trim();
+
+					if(delimiter !== ';'){
+						delimiters.push(delimiter);
+					}
+				}
+			}
+
+			return delimiters;
+		};
+
+		const escapeRegex = function(string) {
+			return string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+		};
+
+		const removeDelimiters = function(queries, delimiter){
+			const removed = [];
+			for(const item of queries){
+				const regex = new RegExp(
+					'DELIMITER',
+					'gim'
+				);
+
+				const trimmed = item
+					.replace(regex, '')
+					.split(delimiter)
+					.filter(text => text.trim().length > 0);
+
+				const length = trimmed.length;
+				for(let idx = 0; idx < length; idx++){
+					removed.push(self._changeDefiner(trimmed[idx]));
+				}
+			}
+
+			return removed;
+		};
+
+		const tmp = [];
+		for(const delimiter of getDelimiters(sql)){
+			const escaped = escapeRegex(delimiter);
+			const regex = new RegExp('DELIMITER\\s' + escaped
+				+ '([^]*?)' + 'DELIMITER', 'gim');
+			const matches = sql.match(regex);
+			const trimmed = removeDelimiters(matches, delimiter);
+
+			tmp.push(...trimmed);
+		}
+
+		for(const query of tmp){
+			if(query.match(/\sPROCEDURE\s/i)) {
+				results.procedures.push(query);
+			} else if(query.match(/\FUNCTION\s/i)) {
+				results.functions.push(query);
+			} else if (query.match(/\sTRIGGER\s/i)) {
+				results.triggers.push(query);
+			}
+		}
+
+		return results;
+	}
+
+	/**
+	 * Get table's name from query
+	 * @private
+	 * @params {string} type - query's type which table's name will be extracted
+	 * @params {string} query - sql query
+	 * @returns {string|null} tables - extracted table's name if presents
+	 * */
+	_getNameFromQuery(type, query){
+		const self = this;
+
+		if(typeof(query) != 'string'){
+			return false;
+		}
+
+		type = type.toUpperCase();
+		let match = null;
+
+		switch(type){
+		case 'ALTER':
+			match = query.match(/ALTER[\s]+TABLE[\s]+(.*?)[\s\(]/im);
+			break;
+
+		case 'TABLE':
+			match = query.match(/CREATE.*?TABLE\s+(?:IF.*?NOT.*?EXISTS){0,}.*?([\w]+)/im);
+			break;
+
+		case 'VIEW':
+			match = query.match(/CREATE.*?VIEW[\s]+(.*?)[\s\(]/im);
+			break;
+
+		case 'PROCEDURE':
+		case 'FUNCTION':
+		case 'ROUTINE':
+			match = query.match(/CREATE.*?(?:PROCEDURE|FUNCTION)[\s]+(.*?)[\s\(]/im);
+			break;
+
+		case 'INSERT':
+			match = query.match(/INSERT\s+INTO[\s]+(.*?)[\s\(]+/im);
+			break;
+
+		default:
+			break;
+		}
+
+		return match ? match[1].replace(/[^0-9a-zA-Z\$\_]/g, '') : null;
+	};
+
+	/**
+	 * Get table's name from queries
+	 * @private
+	 * @params {string} type - query's type which table's name will be extracted
+	 * @params {string[]} queries - array of sql queries with same type
+	 * @returns {string[]} tables - extracted table's names if presents
+	 * */
+	_getNameFromQueries(type, queries){
+		const self = this;
+
+		const names = [];
+
+		for(const query of queries){
+			const name = self._getNameFromQuery(type, query);
+
+			if(name){
+				names.push(name);
+			}
+		}
+
+		return names;
+	};
+
+	/**
+	 * Convert INSERT statements with multiple values into multiple INSERT statements
+	 * @private
+	 * @params {string} query - INSERT statements with multiple values
+	 * @returns {string[]} inserts - multiple INSERT statements
+	 * */
+	_insertMultipleToSingle(query){
+		const self = this;
+		const inserts = [];
+
+		// split values inside query into array. max only 2 nested parentheses.
+		// https://stackoverflow.com/a/35271017/3258981
+		const matches = query.match(/\((?:[^)(]+|\((?:[^)(]+|\([^)(]*\))*\))*\)/gim);
+		if(!matches) return inserts;
+
+		const table = self._getNameFromQuery('INSERT', query);
+		const columns = query.match(/(\(.*?\))\s+VALUES\s+/mi);
+
+		// if columns clause is found, shift the matches as it contains column names
+		if(columns) matches.shift();
+
+		for(const val of matches){
+			inserts.push('INSERT INTO `' + table + '` '
+				+ (columns ? columns[1] : '')
+				+ ' VALUES ' + val + ';');
+		}
+
+		return inserts;
+	}
+
+	/**
+	 * Convert ALTER statements with multiple values
+	 * into multiple ALTER statements for each values
 	 *
-	 * @async
-	 * @param {boolean} [query=true] - should connection be closed after the process is done
+	 * The idea is if 1 query is failed, it won't affect another queries.
+	 * Currently, only support ADD and MODIFY
+	 *
+	 * @private
+	 * @params {string} query - ALTER statements with multiple values
+	 * @returns {string[]} inserts - multiple ALTER statements
+	 * */
+	_alterMultipleToSingle(query){
+		const self = this;
+
+		// count how many times ADD or MODIFY appears in query
+		const count = query.match(/(?:ADD|MODIFY)\s{0,}/gmi);
+
+		// if there's only 1 appearance, return the same one
+		if (!count || count?.length <= 1) return [ query ];
+
+		const table = self._getNameFromQuery('ALTER', query);
+
+		// regex components. It's too long for comfort
+		const reg_REFERENCES = '(?:REFERENCES.*?\\(.*?\\)(?:\\s{1,}ON.*?(?:[\\w|\\s]+){0,}){0,}){0,}';
+		const reg_USING = '(?:USING.*?(?:[\\w|\\s]+))';
+		const reg_REFUSING = `(?:\\s{1,}${reg_REFERENCES}|${reg_USING}){0,}`;
+		const reg_ADDMOD = `(?![^a-z])((?:ADD|MODIFY)\\s{0,}.*?\\(.*?\\)${reg_REFUSING})`;
+
+		const regex = new RegExp(reg_ADDMOD, 'gmi');
+
+		const matches = query.matchAll(regex);
+		const alters = [];
+
+		for (const match of matches) {
+			alters.push('ALTER TABLE `' + table + '` ' + match[1]);
+		}
+
+		return alters;
+	}
+
+	/**
+	 * GET all column's name from CREATE statements
+	 * @private
+	 * @params {string} query - INSERT statements with multiple values
+	 * @returns {string[]} columns - list of column's name
+	 * */
+	_getColumnsFromCreateTable(query){
+
+		const self = this;
+		const columns = {};
+
+		// split values inside query into array. max only 2 nested parentheses.
+		// https://stackoverflow.com/a/35271017/3258981
+		const matches = query.match(/\((?:[^)(]+|\((?:[^)(]+|\([^)(]*\))*\))*\)/gim);
+		if(!matches) return columns;
+
+		const rawCols = matches[0].matchAll(/`(.*?)`\s(.*?)\n/gm);
+
+		for(let col of rawCols){
+			columns[col[1]] = col[2].replace(",","");
+		}
+
+		return columns;
+	};
+
+	/**
+	 * GET all routine's name from FUNCTION, PROCEDURE AND TRIGGER statement
+	 * @private
+	 * @params {string} query - FUNCTION, PROCEDURE AND TRIGGER statements
+	 * @returns {string} name - name of each statement
+	 * */
+	_parseQueryRoutine(query){
+		const self = this;
+
+		try{
+			const parsedQuery = query.match(/\s(FUNCTION|PROCEDURE|TRIGGER)\s`(.*?)`/im);
+			return parsedQuery[2];
+		}catch(err){
+			console.log(err);
+			return "";
+		}
+	}
+
+	/**
+	 * Check if alter statement contains "PRIMARY KEY"
+	 * @private
+	 * @params {string} query - ALTER statements
+	 * @returns {string} name - name of each statement
+	 * */
+	_checkAlterHasPrimary(query){
+		const self = this;
+
+		try{
+			let parsedQuery = query.includes("ADD PRIMARY KEY");
+			if(!parsedQuery){
+				return false;
+			}
+
+			parsedQuery = query.match(/ALTER TABLE `(.*?)`/im);
+
+			return parsedQuery[1] || false;
+		}catch(err){
+			console.log(err);
+			return false;
+		}
+	}
+
+	/**
+	 * Move column definitions inside CREATE TABLE to alter
+	 * @private
+	 * @params {string} query - CREATE TABLE statements
+	 * @returns {{query: string, alter: string[]}} stripped query and list of found ALTER query
+	 * */
+	_stripColumnDefinitionFromCreate(query) {
+		const self = this;
+
+		// remove whitespaces at the beginning of each line
+		query = query.replace(/\n\s+/gmi, '\n');
+
+		const regex = /^((PRIMARY KEY|UNIQUE KEY|KEY|CONSTRAINT)\s+(.*?)(?:(?:,$)|$))/gmi;
+
+		const table = self._getNameFromQuery('TABLE', query);
+		const matches = query.matchAll(regex);
+
+		const alter = [];
+
+		for(const match of matches) {
+			query = query
+				// remove column definition
+				.replace(match[1], '')
+				// remove empty lines
+				.replace(/^\n|\r/gmi, '');
+
+			// add column definitions into alter
+			alter.push(`ALTER TABLE ${table} ADD ${match[2]} ${match[3]}`);
+		}
+
+		return { query, alter };
+	}
+
+	/**
+	 * Remove tables and routines from connected database
+	 * @params {boolean} [query=true] - should connection be closed after the process is done
 	 * */
 	async emptyDatabase(closeConnection = true){
 		const self = this;
@@ -322,8 +586,8 @@ class SQL_IMPORTER {
 		}
 
 		// drop all tables and routines first
-		await self._dropAllTables(self._config.database);
-		await self._dropRoutines(self._config.database);
+		await self.dropAllTables(self._config.database);
+		await self.dropRoutines(self._config.database);
 
 		if(closeConnection){
 			self.close();
@@ -331,9 +595,93 @@ class SQL_IMPORTER {
 	}
 
 	/**
+	 * Get tables relationship indentified by FOREIGN KEYs
+	 * @params {string} query - ALTER query
+	 * @returns {Object} res - tables dependencies
+	 * @returns {string} res.table - table's name
+	 * @returns {string[]} res.dependencies - table's dependencies
+	 * */
+	getForeignKeyRelationship(query){
+		const self = this;
+
+		const table = self._getNameFromQuery('ALTER', query);
+		const FKs = query.match(/FOREIGN\s+KEY\s+.*?\s+REFERENCES\s+(.*?)[\s\(]/gmi);
+
+		const tDep = {
+			table,
+			dependencies: [],
+		};
+
+		if(!FKs){
+			return tDep;
+		}
+
+		const sets = new Set();
+		for(const fk of FKs){
+			const tRelated = fk.match(/FOREIGN\s+KEY\s+.*?\s+REFERENCES\s+(.*?)[\s\(]/i);
+
+			if(!tRelated) continue;
+
+			sets.add(tRelated[1].replace(/[^0-9a-zA-Z\$\_]/g, ''));
+		}
+
+		tDep.dependencies = Array.from(sets);
+
+		return tDep;
+	}
+
+	/**
+	 * Get tables relationship from multiple ALTER queries
+	 * @params {string} queries - ALTER queries
+	 * @returns {Object} res - tables dependencies with table's name as key
+	 * */
+	getTableDependencies(queries){
+		const self = this;
+
+		// change to object to make it easier when trying to find dependencies
+		const tObjDep = {};
+		for(const query of queries){
+			const tDep = self.getForeignKeyRelationship(query);
+			tObjDep[tDep.table] = tDep.dependencies;
+		}
+
+		return tObjDep;
+	}
+
+	/**
+	 * Sort tables by dependencies
+	 * https://stackoverflow.com/a/54346588/3258981
+	 * @params {string} queries - ALTER queries
+	 * @returns {string[]} res - sorted table from less or no dependencies first
+	 * */
+	topologySortDependencies(queries){
+		const self = this;
+
+		const tDep = self.getTableDependencies(queries);
+		const used = new Set;
+		const results = [];
+		let keys = Object.keys(tDep);
+		let i, items, length;
+
+		do {
+			length = keys.length;
+			items = [];
+			keys = keys.filter(k => {
+				if (!tDep[k].every(Set.prototype.has, used)) return true;
+				items.push(k);
+			});
+			results.push(...items);
+			items.forEach(Set.prototype.add, used);
+		} while (keys.length && keys.length !== length);
+
+		results.push(...keys);
+
+		return results;
+	}
+
+	/**
 	 * Load sql file content
-	 *
-	 * @param filepath `string` - SQL file's path
+	 * @params {string} SQL file's path
 	 * */
 	read(filepath){
 		const self = this;
@@ -349,8 +697,7 @@ class SQL_IMPORTER {
 
 	/**
 	 * set SQL content from string
-	 *
-	 * @param contents **[string]** - String containing SQL queries
+	 * @params {string} String containing SQL queries
 	 * */
 	sql(contents){
 		const self = this;
@@ -365,44 +712,109 @@ class SQL_IMPORTER {
 	}
 
 	/**
-	 * Parse sql file. Either #read or #sql must be called before.
-	 *
-	 * @async
-	 * @return res {#ParsedQuery} - parsed sql file
+	 * Parse sql file
+	 * @params {string} filepath - sql file's path
+	 * @returns {Object} res - parsed sql file
+	 * @returns {Object} res.queries - consists of sql queries
+	 * @returns {Object} res.names - consists of tables and routines names
+	 * @returns {string[]} res.queries.functions - consists of FUNCTION CREATE queries
+	 * @returns {string[]} res.queries.procedures - consists of PROCEDURE CREATE queries
+	 * @returns {string[]} res.queries.triggers - consists of TRIGGER CREATE queries
+	 * @returns {string[]} res.queries.table - consists of TABLE CREATE queries
+	 * @returns {string[]} res.queries.alter - consists of ALTER TABLE queries
+	 * @returns {string[]} res.queries.view - consists of CREATE VIEW queries
+	 * @returns {Object} res.queries.inserts - consists of INSERT queries, with each query grouped by table's name as key
+	 * @returns {string[]} res.queries.drop - consists of DROP queries
+	 * @returns {string[]} res.queries.sort - sorted table names based on table's dependencies
+	 * @returns {string[]} res.queries.misc - other queries which not correctly parsed
 	 * */
 	async parse(){
 		const self = this;
 
-		if(!self._contents) throw 'SQL File is not loaded yet!';
+		if(self._contents === undefined){
+			throw 'SQL File is not loaded yet!';
+		}
 
-		const parsed = parseSQLInput(self._contents, self._config);
+		const sql = self._contents;
 
-		console.log('parsed', parsed);
+		if(sql.error){
+			throw 'SQL File is not loaded yet!';
+		}
+
+		const _queries = sql.split(';');
+
+		const length = _queries.length;
+		const queries = {
+			...self.getNonTableCreates(sql),
+			table: [],
+			alter: [],
+			view: [],
+			insert: [],
+			drop: [],
+			sort: [],
+			misc: [],
+		};
+
+		for(let idx = 0; idx < length; idx++){
+			const query = _queries[idx].trim();
+
+			if(query.match(/^INSERT\s/i)){
+				const name = self._getNameFromQuery('INSERT', query);
+
+				if(queries.insert[name] === undefined){
+					queries.insert[name] = [];
+				}
+
+				queries.insert[name].push(query);
+			} else if(query.match(/^ALTER\s/i)){
+				const multiple = self._alterMultipleToSingle(query);
+				queries.alter.push(...multiple);
+			} else if(query.match(/^CREATE\s/i)){
+				if(query.match(/\sVIEW\s/i)){
+					queries.view.push(self._changeDefiner(query));
+				} else if(query.match(/\sTABLE\s/i)){
+					const name = self._getNameFromQuery('TABLE', query);
+
+					if(queries.table[name] === undefined){
+						queries.table[name] = [];
+					}
+
+					// strip col definitions from CREATE TABLE and add them into alter
+					const stripped = self._stripColumnDefinitionFromCreate(query);
+					queries.alter.push(...stripped.alter);
+					queries.table[name].push(stripped.query);
+				}
+			} else if(query.match(/^(DROP\s)/i)){
+				queries.drop.push(query);
+			} else {
+				queries.misc.push(query);
+			}
+		}
+
+		const names = {
+			tables: self._getNameFromQueries('TABLE', queries.table),
+			views: self._getNameFromQueries('VIEW', queries.view),
+			functions: self._getNameFromQueries('FUNCTION', queries.functions),
+			procedures: self._getNameFromQueries('ROUTINE', queries.procedures),
+		};
+
+		queries.sort = self.topologySortDependencies(queries.alter);
 
 		self._contents = undefined;
 
-		return parsed;
+		return { queries, names };
 	}
 
-	/**
-	 * Process table comparation result.
-	 *
-	 * @async
-	 * @private
-	 * @param {ParsedQuery} queries - Parsed SQL contents.
-	 * @param {TableCompare} comparation - Table comparation result.
-	 * */
-	async _processSameTables(queries, comparation) {
+	async _processSameTables(queries, sameTables) {
 		const self = this;
 
 		// get existing routines and trigger
-		const existingRoutine = await self._mysql.query(
-				`SELECT
-					routine_name as name
-				FROM information_schema.routines
-				WHERE routine_schema = ?`,
-				[ self._config.database ]
-			)
+		const existingRoutine = await self._mysql.query(`
+			SELECT
+				routine_name as name
+			FROM information_schema.routines
+			WHERE routine_schema = ?
+		`, [ self._config.database ])
 			.then(res => {
 				const tmp = {};
 				for(let i = 0; i < res.length; i++){
@@ -412,13 +824,12 @@ class SQL_IMPORTER {
 			})
 			.catch(err => {});
 
-		const existingTrigger = await self._mysql.query(
-				`SELECT
-					trigger_name as name
-				FROM information_schema.triggers
-				WHERE trigger_schema = ?`,
-				[ self._config.database ]
-			)
+		const existingTrigger = await self._mysql.query(`
+			SELECT
+				trigger_name as name
+			FROM information_schema.triggers
+			WHERE trigger_schema = ?
+		`, [ self._config.database ])
 			.then(res => {
 				const tmp = {};
 				for(let i = 0; i < res.length; i++){
@@ -428,55 +839,52 @@ class SQL_IMPORTER {
 			})
 			.catch(err => {});
 
-		// FIXME: What if new procedures has different contents?
 		// check duplicate routines and trigger, if it exist then remove new routines
 		for(let i = 0; i < queries.procedures.length; i++){
 			const query = queries.procedures[i];
-			const name = parseQueryRoutine(query);
+			const name = self._parseQueryRoutine(query);
 
 			if(existingRoutine[name]){
 				queries.procedures.splice(i, 1);
 			}
 		}
 
-		// FIXME: What if new functions has different contents?
 		for(let i = 0; i < queries.functions.length; i++){
 			const query = queries.functions[i];
-			const name = parseQueryRoutine(query);
+			const name = self._parseQueryRoutine(query);
 
 			if(existingRoutine[name]){
 				queries.functions.splice(i, 1);
 			}
 		}
 
-		// FIXME: What if new triggers has different contents?
 		for(let i = 0; i < queries.triggers.length; i++){
 			const query = queries.triggers[i];
-			const name = parseQueryRoutine(query);
+			const name = self._parseQueryRoutine(query);
 
 			if(existingTrigger[name]){
 				queries.triggers.splice(i, 1);
 			}
 		}
 
-		// remove alter primary_key in the same table
+		// remove alter primary_key with same table
 		let alterName = {};
 		for(let i = 0; i < queries.alter.length; i++){
-			const name = checkAlterHasPrimary(queries.alter[i]);
-			if(name) alterName[name] = i;
+			const name = self._checkAlterHasPrimary(queries.alter[i]);
+			if(name){
+				alterName[name] = i;
+			}
 		}
 
-		for(let table in comparation){
-			/*
-			 * if existing table:
-			 *
-			 * - has new columns,
-			 *   then remove "create table" and "insert"
-			 *   & replace it with "alter table"
-			 *
-			 * - has no new columns
-			 *   then only remove "create table" and "insert"
-			 * */
+		for(let table in sameTables){
+
+			// for existing table,
+			// but has new columns, then
+			// remove "create table" and "insert"  & replace it with "alter table"
+
+			// for existing table,
+			// but no new columns, then
+			// only remove "create table" and "insert"
 			delete queries.table[table];
 			delete queries.insert[table];
 
@@ -485,8 +893,8 @@ class SQL_IMPORTER {
 				queries.alter.splice(alterName[table], 1);
 			}
 
-			if(comparation[table].new.length > 0){
-				for(const column of comparation[table].new){
+			if(sameTables[table].length > 0){
+				for(let column of sameTables[table]){
 					queries.alter.unshift(
 						"ALTER TABLE `"
 						+ table
@@ -500,181 +908,74 @@ class SQL_IMPORTER {
 		}
 	}
 
-	/**
-	 * Compare tables from provided queries with exisiting ones inside database.
-	 *
-	 * @async
-	 * @private
-	 * @param {Queries} queries - Parsed SQL content.
-	 * @returns {TableCompare} Table Comparation result.
-	 * */
 	async _compareTables(queries) {
 		const self = this;
 
 		const sameTables = {};
 
 		// get existing table & columns
-		const existing = await self._mysql.query(
-				`SELECT
-					table_name,
-					column_name,
-					column_key,
-					column_type,
-					data_type,
-					character_maximum_length,
-					is_nullable,
-					column_default
-				FROM information_schema.COLUMNS
-				WHERE TABLE_SCHEMA = ?`,
-				[ self._config.database ]
-			)
+		const existingTables = await self._mysql.query(`
+			SELECT table_name, column_name
+			FROM information_schema.COLUMNS
+			WHERE TABLE_SCHEMA = ?
+		`,[self._config.database])
 			.then(res => {
+
 				const tables = {};
 
 				// grouping columns by it's table
-				const tblLen = res.length;
-				for(let i = 0; i < tblLen; i++){
+				for(let i = 0; i < res.length; i++){
 
-					if (!res[i]) continue;
+					const curTbl = res[i]?.table_name;
+					const curCol = res[i]?.column_name;
 
-					const current = {
-						table: res[i].table_name,
-						column: res[i].column_name,
-					};
+					if(!(curTbl in tables)){
+						tables[curTbl] = {};
+					}
 
-					if(!tables[current.table]) tables[current.table] = {};
+					const newObj = {};
+					newObj[curCol] = true;
 
-					const newObj = {
-						[current.column]: {
-							key: res[i].column_key || undefined,
-							type: res[i].column_type,
-							datatype: res[i].data_type,
-							length: res[i].data_type === 'varchar'
-								? Number(res[i].character_maximum_length)
-								: 0,
-							is_nullable: res[i].is_nullable.toUpperCase() === 'YES',
-							default: res[i].column_default ?? undefined,
-						},
-					};
-
-					Object.assign(tables[current.table], newObj);
+					Object.assign(tables[curTbl], newObj);
 				}
 
 				return tables;
 			})
 			.catch(err => {
-				console.error('_compareTables', err);
+				console.log(err);
 				return undefined;
 			});
 
 		// no need to go any further
-		if (existing === undefined) return undefined;
-
-		const list = [];
+		if (existingTables === undefined) return undefined;
 
 		// compare every existing tables
-		for(const tbl of queries.sort){
-			if(!existing[tbl] || !queries.table[tbl]) continue;
+		for(let tbl of queries.sort){
+			if(existingTables[tbl] && queries.table[tbl]){
 
-			// compare every columns in each tables
-			const query = queries.table[tbl][0];
-			const requested = getColumnsFromCreateTable(query);
-
-			list[tbl] = {
-				new: [],
-				same: [],
-				mod: [],
-				nomore: Object.keys(existing[tbl]),
-			};
-
-			for(const reqCol in requested){
-				const detail = requested[reqCol];
-
-				const now = parseColumnCreateQuery(detail);
-				const old = existing[tbl][reqCol];
-
-				// column doesn't exist
-				if (old === undefined) {
-					list[tbl].new.push({ name: reqCol, detail });
-					continue;
+				sameTables[tbl] = [];
+				// compare every columns in each tables
+				const columns = self._getColumnsFromCreateTable(queries.table[tbl][0]);
+				for(let column in columns){
+					if(!(column in existingTables[tbl])){
+						sameTables[tbl].push({
+							name: column,
+							detail: columns[column]
+						});
+					}
 				}
 
-				// remove from no_more
-				const idx = list[tbl].nomore.indexOf(reqCol);
-				list[tbl].nomore[idx] = undefined;
-
-				if (isColumnIdentical(now, old)) {
-					list[tbl].same.push({ name: reqCol, detail });
-				} else {
-					list[tbl].mod.push({ name: reqCol, detail });
-				}
 			}
-
-			// remove undefine element fro array
-			list[tbl].nomore = list[tbl].nomore.filter(x => x);
 		}
 
-		return list;
+		return sameTables;
 	}
 
 	/**
-	 * Execute queries.
-	 *
-	 * @private
-	 * @param queries **string[]** List of SQL Queries.
-	 * @param verbose **number** Verbosity level.
-	 * @returns {FailedQuery[]} List of failed queries.
-	 * */
-	async _execute(queries, verbose){
-		const self = this;
-
-		const _failedQueries= [];
-
-		for(const query of queries){
-			try{
-				const res = await self._mysql.query(
-					query.match(/^INSERT\s/)
-						? query
-						: 'SET FOREIGN_KEY_CHECKS=0;' + query
-				);
-
-				if(verbose > 1){
-					console.log(
-						'\n================= QUERY ===================\n',
-						query,
-						'\n===========================================\n',
-						res,
-						'\n================ SUCCESS ==================\n',
-					);
-				}
-			} catch(error) {
-				if(verbose > 0){
-					console.error(
-						'\n+++++++++++++++++ QUERY +++++++++++++++++++\n',
-						query,
-						'\n+++++++++++++++++ +++++ +++++++++++++++++++\n',
-						error,
-						'\n+++++++++++++++++ ERROR +++++++++++++++++++\n',
-					);
-				}
-
-				_failedQueries.push({
-					code: error.errno,
-					msg: error.sqlMessage,
-					query,
-				});
-			}
-		}
-
-		return _failedQueries;
-	};
-
-	/**
 	 * Import sql file into connected database
-	 *
-	 * @param {Object} options - options
-	 * @param {boolean} [options.dropFirst=true] - whether perform wipe database first before import or not
-	 * @param {boolean|string} [options.withData=true] - whether perform insert data or not. Set to 'single' to insert data individually
+	 * @params {Object} options - options
+	 * @params {boolean} [options.dropFirst=true] - whether perform wipe database first before import or not
+	 * @params {boolean|string} [options.withData=true] - whether perform insert data or not. Set to 'single' to insert data individually
 	 * @returns {string[]} failed - failed sql queries
 	 * @returns {number} failed.code - sql error code
 	 * @returns {string} failed.query - failed sql query
@@ -689,10 +990,60 @@ class SQL_IMPORTER {
 			});
 		}
 
-		let { withData, dropFirst, closeConnection } = options;
+		let { withData, dropFirst, closeConnection, compareExisting } = options;
 
-		if (withData === undefined) withData = true;
-		if (dropFirst === undefined) dropFirst = true;
+		if(withData === undefined){
+			withData = true;
+		}
+
+		if(dropFirst === undefined){
+			dropFirst = true;
+		}
+
+		if(compareExisting === undefined){
+			compareExisting = false;
+		}
+
+		const execAll = async function(queries, verbose){
+			const _failedQueries= [];
+
+			for(const query of queries){
+				try{
+					const res = await self._mysql.query(
+						query.match(/^INSERT\s/)
+							? query
+							: 'SET FOREIGN_KEY_CHECKS=0;' + query
+					);
+
+					if(verbose > 1){
+						console.log(
+							'\n================= QUERY ===================\n',
+							query,
+							'\n===========================================\n',
+							res,
+							'\n================ SUCCESS ==================\n',
+						);
+					}
+				} catch(error) {
+					if(verbose > 0){
+						console.error(
+							'\n+++++++++++++++++ QUERY +++++++++++++++++++\n',
+							query,
+							'\n+++++++++++++++++ +++++ +++++++++++++++++++\n',
+							error,
+							'\n+++++++++++++++++ ERROR +++++++++++++++++++\n',
+						);
+					}
+
+					_failedQueries.push({
+						code: error.errno,
+						query,
+					});
+				}
+			}
+
+			return _failedQueries;
+		};
 
 		if(dropFirst){
 			// drop all tables and routines first
@@ -711,14 +1062,15 @@ class SQL_IMPORTER {
 		const { verbose } = self._config;
 		const _failed = [];
 
-		if(dropFirst){
-			// execute drop queries if any
-			_failed.push(... await self._execute(queries.drop, verbose));
-		} else {
-			// compare existing table & column
-			const comparation = await self._compareTables(queries);
-			self._processSameTables(queries, comparation);
+
+		// compare existing table & column
+		if(compareExisting){
+			const sameTables = await self._compareTables(queries);
+			if(sameTables) self._processSameTables(queries, sameTables);
 		}
+
+		// create db structure first
+		_failed.push(... await execAll(queries.drop, verbose));
 
 		// create tables first
 		for(const tbl of queries.sort){
@@ -738,24 +1090,24 @@ class SQL_IMPORTER {
 				}
 			}
 
-			_failed.push(... await self._execute(tQueries, verbose));
+			_failed.push(... await execAll(tQueries, verbose));
 		}
 
 		// then create routine
-		_failed.push(... await self._execute(queries.functions, verbose));
-		_failed.push(... await self._execute(queries.procedures, verbose));
+		_failed.push(... await execAll(queries.functions, verbose));
+		_failed.push(... await execAll(queries.procedures, verbose));
 
 		// alter tables and create trigger at the end
-		_failed.push(... await self._execute(queries.alter, verbose));
-		_failed.push(... await self._execute(queries.triggers, verbose));
+		_failed.push(... await execAll(queries.alter, verbose));
+		_failed.push(... await execAll(queries.triggers, verbose));
 
 		// exported sql from phpmyadmin might already add view as table
 		// drop them first
 		for(const view of names.views){
-			await self._dropTable(view);
+			await self.dropTable(view);
 		}
 
-		_failed.push(... await self._execute(queries.view, verbose));
+		_failed.push(... await execAll(queries.view, verbose));
 
 		// add data
 		if(withData){
@@ -776,16 +1128,15 @@ class SQL_IMPORTER {
 				switch(withData){
 				case 1:
 				case true:
-					_failed.push(... await self._execute(tQueries, verbose));
+					_failed.push(... await execAll(tQueries, verbose));
 					break;
 
 				case 2:
 				case 'SINGLE':
 				case 'single':
 					for(const item of tQueries){
-						const item = tQueries[idx];
-						const singles = insertMultipleToSingle(item);
-						_failed.push(... await self._execute(singles, verbose));
+						const singles = self._insertMultipleToSingle(item);
+						_failed.push(... await execAll(singles, verbose));
 					}
 					break;
 
@@ -799,9 +1150,9 @@ class SQL_IMPORTER {
 
 		return {
 			status: _failed.length <= 0,
-			error:  _failed.length <= 0 ? undefined : _failed,
+			error: _failed,
 		};
 	}
 }
 
-module.exports = SQL_IMPORTER;
+module.exports = new IMPORTER();
